Cache QA's label on ingest
Adds new class to override built-in form-to-attributes translator,
caching labels as the "hidden" label.  Hacks a quick bandaid fix to
RdfResource to ensure labels can be removed - required for testing
this change, but also something we need to figure out properly long-
term.

Fixes multiple tests to stop sending invalid data to the subject field,
now that subjects are validated by virtue of trying to cache their
label.