Fix property name
Not sure all the magic involved, but a property turns into a method
somewhere that also has RDF methods.  Collisions are not reported,
rather they are silently skipped, so attempts to get the :repository
property don't return said property; they return the RDF repository,
which causes all kinds of not-so-obvious problems.